Update patch set 1
Patch Set 1:

Yes indeed, most CM devices should work with SurfaceFLinger to get a screenshot and GPU-less devices usually aren't BGRA. 

As I explained here: http://code.google.com/p/cyanogenmod/issues/detail?id=4446 I'm a developer at the Replicant project, that is a 100% free CyanogenMod fork. As a result, we don't include the video acceleration non-free stuff so the screenshot is done via framebuffer. On the Nexus S (upcoming supported device on Replicant), that has BGRA framebuffer, we need my patches to have a screenshot with the correct colors.

This doesn't particularly affect CM concretely but I though it was a good idea to send the patch upstream (so here, to CM) because:
* it makes the CMScreenshot app more complete
* it doesn't require me to do a fork of it to apply my fix

Though, if you don't want to include this, I'll fork and apply my patch on Replicant repos.

Patch-set: 1